backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==7.0.0
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==2.4.0
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==39.0.2
daphne==2.5.0
defusedxml==0.7.1
Django==3.2
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==22.1
django-graphql-jwt==0.3.4
django-graphql-social-auth==0.1.4
django-taggit==3.1.0
djongo==1.3.6
dnspython==2.3.0
frozenlist==1.3.3
graphene==2.1.9
graphene-django==2.15.0
graphene-subscriptions==1.0.2
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
multidict==6.0.4
oauthlib==3.2.2
promise==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pymongo==3.12.3
pyOpenSSL==23.1.1
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
Rx==1.6.3
service-identity==21.1.0
singledispatch==4.0.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.2.4
text-unidecode==1.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
yarl==1.8.2
zope.interface==6.0
